+++ CMakeLists.txt
@@ -92,7 +92,7 @@ set(FISH_BUILTIN_SRCS
     src/builtins/disown.cpp src/builtins/echo.cpp src/builtins/emit.cpp
     src/builtins/eval.cpp src/builtins/exit.cpp src/builtins/fg.cpp
     src/builtins/function.cpp src/builtins/functions.cpp src/builtins/history.cpp
-    src/builtins/jobs.cpp src/builtins/math.cpp src/builtins/printf.cpp src/builtin_path.cpp
+    src/builtins/jobs.cpp src/builtins/math.cpp src/builtins/printf.cpp src/builtins/path.cpp
     src/builtins/pwd.cpp src/builtins/random.cpp src/builtins/read.cpp
     src/builtins/realpath.cpp src/builtins/return.cpp src/builtins/set.cpp
     src/builtins/set_color.cpp src/builtins/source.cpp src/builtins/status.cpp
