+++ src/builtin.cpp
@@ -50,6 +50,7 @@
 #include "builtins/history.h"
 #include "builtins/jobs.h"
 #include "builtins/math.h"
+#include "builtins/path.h"
 #include "builtins/printf.h"
 #include "builtins/pwd.h"
 #include "builtins/random.h"
@@ -65,7 +66,6 @@
 #include "builtins/type.h"
 #include "builtins/ulimit.h"
 #include "builtins/wait.h"
-#include "builtin_path.h"
 #include "common.h"
 #include "complete.h"
 #include "exec.h"
